refactor main flow
